[OPPRO-153] Fix int variant (#17)

+++ velox/substrait/TypeUtils.h
@@ -36,7 +36,7 @@ template <>
 struct RangeTraits<TypeKind::INTEGER> {
   using RangeType = common::BigintRange;
   using MultiRangeType = common::BigintMultiRange;
-  using NativeType = int64_t;
+  using NativeType = int32_t;
 };
 
 template <>
